Replaced cert tab with inline DNS toggle
